ARC-AGI is now up and running via docker
- matched standard Flask docker boilerplate
   - moved the JS and CSS files into /static
   - moved the HTML file to /templates, renamed testing_interface.html to index.html
   - modified app.py to render_templates (so that the Flask server spits out to the HTML automatically)
   - modified the HTML file to correctly reference the static JS and CSS files when running in a container
